Unify and simplify c++ tests
