--- cesiumOmniverseCppTestsExtension/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_subdirectory(bindings)
-add_subdirectory(public)
