--- cesiumOmniverseCppTestsExtension/include/UsdUtilTests.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void runAllUsdUtilTests();
