+++ tests/public/CMakeLists.txt
@@ -4,6 +4,10 @@ glob_files(SOURCES "${CMAKE_CURRENT_LIST_DIR}/*.cpp")
 
 get_property(ADDITIONAL_LIBRARIES GLOBAL PROPERTY NVIDIA_ADDITIONAL_LIBRARIES_PROPERTY)
 
+# replace a string in the utils header with the intended working dir for the test executable
+set(TEST_WORKING_DIRECTORY "${PROJECT_SOURCE_DIR}")
+configure_file("${CMAKE_CURRENT_LIST_DIR}/testUtils.h.in" "${CMAKE_CURRENT_LIST_DIR}/../include/testUtils.h")
+
 # cmake-format: off
 setup_lib(
     TARGET_NAME
@@ -14,10 +18,11 @@ setup_lib(
     SOURCES
         ${SOURCES}
     INCLUDE_DIRS
-        "${PROJECT_SOURCE_DIR}/cesiumOmniverseCppTestsExtension/include"
+        "${PROJECT_SOURCE_DIR}/tests/include"
     LIBRARIES
         CesiumOmniverseCore
         doctest::doctest
+        yaml-cpp::yaml-cpp
     ADDITIONAL_LIBRARIES
         # Unfortunately we need this in both cesium.omniverse.plugin and CesiumOmniverseCore because we're bypassing
         # CMake's built-in dependency system
